CRUD permisos Login query to DB

+++ BussinessLayer/UsesCases/UsuarioPermisoController.cs
@@ -30,6 +30,10 @@ namespace BussinessLayer.UsesCases
             {
                 using (Entities db = new Entities())
                 {
+                    var exist = db.UsuarioModulo.Where(el => el.IdModulo.Equals(element.IdModulo) && el.IdUsuario.Equals(element.IdUsuario)).FirstOrDefault();
+                    if (exist != null)
+                        return exist;
+
                     element = db.UsuarioModulo.Add(element);
                     db.SaveChanges();
                 }
@@ -51,6 +55,26 @@ namespace BussinessLayer.UsesCases
                 using (Entities db = new Entities())
                 {
                     UsuarioModulo toDelete = db.UsuarioModulo.Find(id);
+                    db.UsuarioModulo.Remove(toDelete);
+                    db.SaveChanges();
+                }
+                return true;
+            }
+            catch (Exception e)
+            {
+                Log.Write("Ha ocurrido un error " + e.Message);
+            }
+            return false;
+        }
+        public bool DeleteByPermiso(UsuarioModulo element)
+        {
+            if (element == null) return false;
+
+            try
+            {
+                using (Entities db = new Entities())
+                {
+                    var toDelete = db.UsuarioModulo.Where(el => el.IdModulo.Equals(element.IdModulo) && el.IdUsuario.Equals(element.IdUsuario)).FirstOrDefault();
                     db.UsuarioModulo.Remove(toDelete);
                     db.SaveChanges();
                 }
@@ -92,6 +116,34 @@ namespace BussinessLayer.UsesCases
                 {
                     db.Configuration.LazyLoadingEnabled = true;
                     lista = db.UsuarioModulo.Where(el => el.IdUsuario.Equals(IdUsuario)).Select(el => el.Modulo).ToList();
+                }
+                return lista;
+            }
+            catch (Exception e)
+            {
+                string s = e.Message;
+                Log.Write("Ha ocurrido un error " + s);
+            }
+            //Retorna lista vacia para evitar excepciones en llamada
+            return new List<Modulo>();
+        }
+        public List<Modulo> GetListaExcludePermisoByUsuario(int IdUsuario)
+        {
+            List<Modulo> lista = null;
+            try
+            {
+                using (Entities db = new Entities())
+                {
+                    var todas = from p in db.Modulo
+                                select p as Modulo;
+
+                    var asignadas = from p in db.Modulo
+                                    join up in db.UsuarioModulo
+                                    on p.Id equals up.IdModulo
+                                    where up.IdUsuario == IdUsuario
+                                    select p as Modulo;
+
+                    lista = todas.Except(asignadas).ToList();
                 }
                 return lista;
             }
